Add mutableOf and immutableOf value factory methods that return the exact type.
Now mutableOf and immutableOf will actually return the mutable or immutable type, so you don't need to call asMutable or asImmutable. 
Value.mutableOf and Value.immutableOf for BoundedValue's only works for comparable values, because of a java limitation. Use BoundedValue.mutableOf and BoundedValue.immutableOf in those cases. The generic methods are now called genericMutableOf and genericImmutableOf, these still require the asMutable and asImmutable calls for the exact type.